feat(graph): reorder store
