+++ src/graph/graph-store/StoreProvider.tsx
@@ -1,10 +1,11 @@
 import React, { ReactNode } from 'react'
 import createContext from 'zustand/context'
-import { GraphStore, createStore } from './graph-store'
+import { createStore } from './store'
+import { GraphStore } from './model'
 
 const { Provider, useStore } = createContext<GraphStore>()
-
 export const StoreProvider = ({ children }: { children: ReactNode }) => {
   return <Provider createStore={createStore}>{children}</Provider>
 }
+
 export const useGraphStore = useStore
